# Speed up query execution