--- a/dockers/llm.rag.service/common.py
+++ b/dockers/llm.rag.service/common.py
@@ -51,7 +51,6 @@ class State(TypedDict):
     question: str
     query: str
     result: str
-    answer: str
 
 
 class SearchType(Enum):