--- a/dockers/llm.rag.service/common_test.py
+++ b/dockers/llm.rag.service/common_test.py
@@ -1,6 +1,5 @@
 from common import trim_answer
 
-
 test_string = """
 A recent SSH issue customers had was related to SSH access being flagged as a problem in their old environment because it was under a reseller account. This issue was identified as a "Won\'t Fix" and the Linux Support Engineer informed the customer that they would need to let their security vendor know about these details so they can be excluded from the report as "False Positives."<|im_end|>