--- a/dockers/llm.rag.service/proxy_app.py
+++ b/dockers/llm.rag.service/proxy_app.py
@@ -1,4 +1,5 @@
 import os
+
 import uvicorn