We called write_query with the same parameters twice. We can call it once and use the output twice. This should save us approx. 25% of execution time.

Measured before this change:
```
----> Write query execution time: 11.730087 seconds
----> Execute query execution time: 11.444552 seconds
----> Get answer execution time: 20.009238 seconds

----> Execution time: 43.491239 seconds
``` 

measured after this change:
```
----> Write query execution time: 6.158734 seconds
----> Execute query execution time: 0.000908 seconds
----> Get answer execution time: 11.245217 seconds

----> Execution time: 17.741704 seconds
```

Additionally some formatting and import sorting was applied.